pyproject.toml
[tool.poetry]
name = "llama-index-demo"
version = "0.1.0"
description = ""
authors = ["Owen Fraser-Green <owen@fraser-green.com>"]
readme = "README.md"
packages = [{include = "llama_index_demo"}]

[tool.poetry.dependencies]
python = "^3.11"
llama-index = "^0.6.7"
nest-asyncio = "^1.5.6"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
